#100: folder refactoring
