+++ src/net/KNetPS/ClientCmdlet/InvokeSubscribeCmdletCommand.cs
@@ -21,7 +21,7 @@ using Java.Util.Regex;
 using MASES.KNet.Clients.Consumer;
 using System.Management.Automation;
 
-namespace MASES.KNetPS.CodeCmdlet
+namespace MASES.KNetPS.ClientCmdlet
 {
     [Cmdlet(VerbsLifecycle.Invoke, "Subscribe")]
     public class InvokeSubscribeCmdletCommand : PSCmdlet
