+++ src/net/KNetPS/ClientCmdlet/NewProducerConfigBuilderCmdletCommand.cs
@@ -19,7 +19,7 @@
 using MASES.KNet.Clients.Producer;
 using System.Management.Automation;
 
-namespace MASES.KNetPS.CodeCmdlet
+namespace MASES.KNetPS.ClientCmdlet
 {
     [Cmdlet(VerbsCommon.New, "ProducerConfigBuilder")]
     [OutputType(typeof(KNetProducer<,>))]
